Re-enable triangle counting perf test (#1684, #1745)
But disable non-working modes:
- gpu backends
- AI, IA, IAUNION algorithms
