+++ perf_test/graph/CMakeLists.txt
@@ -16,10 +16,8 @@ KOKKOSKERNELS_ADD_EXECUTABLE(
   SOURCES KokkosGraph_mis_d2.cpp       
   )
 
-
-#Below will probably fail on GPUs.
-#KOKKOSKERNELS_ADD_EXECUTABLE(
-#  graph_triangle
-#  SOURCES KokkosGraph_triangle.cpp      
-#  )
+KOKKOSKERNELS_ADD_EXECUTABLE(
+  graph_triangle
+  SOURCES KokkosGraph_triangle.cpp      
+  )
